Move raft-related tests out of store_split_test.go.
Rename all tests in package storage_test.go to client_*_test.go
to make the division clearer.
