+++ storage/client_merge_test.go
